requirements.txt
aiohttp==3.8.3
aiolinebot==0.4.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.92.0
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
idna==3.4
line-bot-sdk==2.4.1
multidict==6.0.4
openai==0.26.5
pydantic==1.10.5
pyngrok==5.2.1
python-decouple==3.7
PyYAML==6.0
requests==2.28.1
sniffio==1.3.0
starlette==0.25.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
yarl==1.8.2
